third_party/WebKit/LayoutTests/inspector/elements/elements-hide-html-comments-expected.txt

Verifies show/hide HTML comments setting.
HTML comments shown:
  <!-- comment 1 -->
- <html>
    + <head>…</head>
      <!-- comment 2 -->
    - <body onload="runTest()">
        - <p>
              "\nVerifies show/hide HTML comments setting.\n"
              <span id="inspect"></span>
              <!-- comment 3 -->
          </p>
      </body>
  </html>

HTML comments hidden:
- <html>
    + <head>…</head>
    - <body onload="runTest()">
        + <p>…</p>
      </body>
  </html>
